Move forge commands under config forge
Move the standalone `but forge` commands into the `but config forge`
namespace and remove the top-level `but forge` command. This centralizes
forge configuration and management under the config command, updates CLI
args and help text, and adjusts command dispatch and metrics to reflect
the new subcommand layout.

- Add ForgeSubcommand enum and subcommand handling to args/config.rs to support `auth`, `list-users`, and `forget` under `but config forge`.
- Update command::config to accept and dispatch ForgeSubcommand, split forge logic into show, auth, list-users, and forget handlers, and update prompts and messages to reference `but config forge`.
- Remove the top-level `Subcommands::Forge` handling and usage; wire config platform dispatch into main binary so `config forge` is used instead.
- Adjust metrics and imports to record forge auth/list/forget events under the config command path.

These changes were needed to unify forge-related functionality under the
configuration command, simplify top-level commands, and provide clearer,
consistent UX for forge authentication and account management.
Show forge overview by default and remove forge command set

Make `but config forge` (overview) run like `list-users` and display
available options, credential status, and re-authentication hints. This
change consolidates the human, shell, and JSON outputs: human output now
lists authenticated GitHub accounts with validity checks and commands
(auth/forget), shell output prints usernames, and JSON output returns
structured account data. Also remove the `forge` command module from the
command registry since the command set is no longer used.
