--- crates/but/src/command/forge/mod.rs
@@ -1,2 +1,0 @@
-mod auth;
-pub mod integration;
